Added more unit tests
