+++ src/converter/core/styles/opacity.spec.ts
@@ -1,0 +1,20 @@
+/**
+ * @license
+ * Copyright Andrey Chalkin <L2jLiga>. All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://github.com/L2jLiga/xd2svg/LICENSE
+ */
+
+import * as assert from 'assert';
+import { opacity } from './opacity';
+
+describe('Core > Styles parsers > Opacity', () => {
+  it('should return url with href', () => {
+    const given = 'block';
+
+    const result = opacity.parse(given);
+
+    assert.equal(result, given);
+  });
+});
