+++ src/converter/core/utils/defs-list.spec.ts
@@ -1,0 +1,8 @@
+import * as assert from 'assert';
+import { defs } from './defs-list';
+
+describe('Definitions list', () => {
+  it('should export definitions list', () => {
+    assert.ok(defs);
+  });
+});
