+++ src/converter/core/utils/gradients-list.spec.ts
@@ -1,0 +1,8 @@
+import * as assert from 'assert';
+import { gradients } from './gradients-list';
+
+describe('Gradients list', () => {
+  it('should export gradients list', () => {
+    assert.ok(gradients);
+  });
+});
